StashApiClient: Remove wrappers around mapper.readValue()
Calling mapper.readValue() directly make the code more readable.

Move a test for parsing StashPullRequestMergeableResponse to a new class,
StashPullRequestMergeableResponseTest. Shorten the example JSON string.
