+++ src/test/java/stashpullrequestbuilder/stashpullrequestbuilder/stash/StashApiClientTest.java
@@ -114,17 +114,6 @@ public class StashApiClientTest {
   }
 
   @Test
-  public void testParsePullRequestMergeStatus() throws Exception {
-    StashPullRequestMergeableResponse resp =
-        StashApiClient.parsePullRequestMergeStatus(
-            "{\"canMerge\":false,\"conflicted\":false,\"vetoes\":[{\"summaryMessage\":\"You may not merge after 6pm on a Friday.\",\"detailedMessage\":\"It is likely that your Blood Alcohol Content (BAC) exceeds the threshold for making sensible decisions regarding pull requests. Please try again on Monday.\"}]}");
-    assertThat(resp, is(notNullValue()));
-    assertThat(resp.getCanMerge(), is(false));
-    assertThat(resp.getConflicted(), is(false));
-    assertThat(resp.getVetoes(), hasSize(1));
-  }
-
-  @Test
   public void getPullRequests_gets_empty_list() throws Exception {
     stubFor(get(pullRequestPath(0)).willReturn(jsonResponse("PullRequestListEmpty.json")));
